+++ debug_toolbar/panels/timer.py
@@ -17,6 +17,8 @@ class TimerPanel(Panel):
     Panel that displays the time a response took in milliseconds.
     """
 
+    is_async = True
+
     def nav_subtitle(self):
         stats = self.get_stats()
         if stats.get("utime"):