+++ docs/changes.rst
@@ -4,6 +4,7 @@ Change log
 Pending
 -------
 
+* Added support for async to timer panel.
 * Added a note about the default password in ``make example``.
 * Removed logging about the toolbar failing to serialize a value into JSON.
 * Moved the the import statement of ``debug_toolbar.urls`` to within the if